RISC-V: memmove size optimized version: use compressed registers only
changed register t1 to register a4, so that the affected instructions
can be compressed. since now we have less registers available, the
following changes needed to be made:
in the previous version of this function, a4 was used to hold the offset
that needs to be added to source and destination addresses before copying
any data in the case of source address > destination address.
since a4 now holds the destination address, this offset is not calculated
anymore. instead, a2 (the number of bytes to be copied) is added to source
and destination addresses. therefore, in the case of source address >
destination adress, a value of 1 needs to be subtracted from both addresses
before starting the copying process.