RISC-V: Optimize memcpy space by rewriting fully in assembly
This new version is based on the previous one but has two main
differences:

- Instead of copying the remainder of 64 bytes one byte at a time, the
  new implementation copies at most one 32-byte block before copying one
  byte at a time.
- In the case where __riscv_misaligned_avoid is defined, the C
  implementation copies one byte at a time if the src and dest
  alignments differ. The new implementation aligns dest to SZREG, does
  SZREG byte-sized loads, packs them in one register, and uses one
  aligned SZREG-sized store.

The new implementation is 18% smaller than the C version compiled with
-O3 with GCC 14.2.

In terms of performance, gem5 was used to emulate a small, in-order core
to exercise all possible pairs of src & dest alignments, and different
copy sizes. Percentages are the average reduction in execution (number
of cycles). Higher is better.

Average over all alignment pairs:
- `-mtune=thead-c906 -mno-strict-align -O3`: 2%
- `-mtune=thead-c906 -mstrict-align -O3`: 13%

Average over different sizes (0, 1, 2, 3, 4, 8, 16, 32, 64, 128, 200,
256, 300, 400, 512, 700, 800, 1000, 1024, 2048, 4096, 8192 16384):
- `-mtune=thead-c906 -mno-strict-align -O3`: 3%
- `-mtune=thead-c906 -mstrict-align -O3`: 7%