RISC-V: Optimize memcpy space by rewriting fully in assembly
This new version is inspired by the implementation for aarch64, but uses
smaller "categories" of sizes (64, 32, 16, and 4 bytes).

Similar to the previous C implementation, this version falls back to
byte-wise copy for source and destination pairs that cannot be aligned
to (xlen / 8) bytes when __riscv_misaligned_avoid is defined.

The new implementation is 15% smaller than the C version compiled with
-O3 with GCC 14.2.

In terms of performance, measured in `spike`, the new implementation is
achieving 1.4% performance improvement on various sizes and alignments
over the previous C (-O3) implementation.