RISC-V: memcpy() align dest when misaligned access is prohibited
Add a code path for when source and dest are differently aligned.

If misaligned access is slow or prohibited, and the alignments of the
source and destination are different, we align the destination to do
XLEN stores. This uses only one aligned store for every four (or eight
for XLEN == 64) bytes of data.