RISC-V: memmove() speed optimized: Align src if at least one address is unaligned
Always align the source address if at least one address is unaligned.
Load one word (or one doubleword for rv64) at a time and store back
individual bytes.
This change decreases the number of retired instructions if at least
one address is unaligned. It also affects the case where both addresses
are aligned, a destructive overlap of the memory areas is present, and
length is not a multiple of SZREG. This is because in the case of a
destructive overlap, the length is added to both addresses before starting
the copying process.